Add public API for creating SvgDecoder.Factory from commonMain
