+++ coil-svg/src/commonMain/kotlin/coil3/decode/SvgDecoder.kt
@@ -1,0 +1,11 @@
+package coil3.decode
+
+object SvgDecoder {
+    fun Factory(
+        useViewBoundsAsIntrinsicSize: Boolean = true
+    ): Decoder.Factory = SvgDecoderFactory(useViewBoundsAsIntrinsicSize)
+}
+
+internal expect fun SvgDecoderFactory(
+    useViewBoundsAsIntrinsicSize: Boolean,
+): Decoder.Factory
