+++ coil-svg/src/nonAndroidMain/kotlin/coil3/decode/SvgDecoder.nonAndroid.kt
@@ -1,23 +1,18 @@
 package coil3.decode
 
-import android.graphics.Canvas
-import android.graphics.RectF
-import androidx.core.graphics.createBitmap
-import androidx.core.graphics.drawable.toDrawable
 import coil3.ImageLoader
 import coil3.asCoilImage
 import coil3.fetch.SourceFetchResult
 import coil3.request.Options
-import coil3.request.bitmapConfig
-import coil3.request.css
 import coil3.size.Scale
 import coil3.size.isOriginal
 import coil3.util.toPx
-import coil3.util.toSoftware
-import com.caverock.androidsvg.RenderOptions
-import com.caverock.androidsvg.SVG
 import kotlin.math.roundToInt
-import kotlinx.coroutines.runInterruptible
+import org.jetbrains.skia.*
+import org.jetbrains.skia.svg.SVGDOM
+
+internal actual fun SvgDecoderFactory(useViewBoundsAsIntrinsicSize: Boolean): Decoder.Factory =
+    SkiaSvgDecoder.Factory(useViewBoundsAsIntrinsicSize)
 
 /**
  * A [Decoder] that uses [AndroidSVG](https://bigbadaboom.github.io/androidsvg/) to decode SVG
@@ -26,24 +21,26 @@ import kotlinx.coroutines.runInterruptible
  * @param useViewBoundsAsIntrinsicSize If true, uses the SVG's view bounds as the intrinsic size for
  *  the SVG. If false, uses the SVG's width/height as the intrinsic size for the SVG.
  */
-class SvgDecoder @JvmOverloads constructor(
+private class SkiaSvgDecoder(
     private val source: ImageSource,
     private val options: Options,
     val useViewBoundsAsIntrinsicSize: Boolean = true,
 ) : Decoder {
 
-    override suspend fun decode() = runInterruptible {
-        val svg = source.source().use { SVG.getFromInputStream(it.inputStream()) }
+    override suspend fun decode(): DecodeResult {
+        val bytes = source.source().readByteArray()
+        val svg = SVGDOM(Data.makeFromBytes(bytes))
 
         val svgWidth: Float
         val svgHeight: Float
-        val viewBox: RectF? = svg.documentViewBox
+        val viewBox: Rect? = svg.root?.viewBox
+
         if (useViewBoundsAsIntrinsicSize && viewBox != null) {
-            svgWidth = viewBox.width()
-            svgHeight = viewBox.height()
+            svgWidth = viewBox.width
+            svgHeight = viewBox.height
         } else {
-            svgWidth = svg.documentWidth
-            svgHeight = svg.documentHeight
+            svgWidth = svg.root?.width?.value ?: 0f
+            svgHeight = svg.root?.height?.value ?: 0f
         }
 
         val bitmapWidth: Int
@@ -66,18 +63,25 @@ class SvgDecoder @JvmOverloads constructor(
 
         // Set the SVG's view box to enable scaling if it is not set.
         if (viewBox == null && svgWidth > 0 && svgHeight > 0) {
-            svg.setDocumentViewBox(0f, 0f, svgWidth, svgHeight)
+            svg.root?.viewBox = Rect.makeWH(svgWidth, svgHeight)
         }
 
-        svg.setDocumentWidth("100%")
-        svg.setDocumentHeight("100%")
+        val bitmap = Bitmap().apply {
+            allocN32Pixels(bitmapWidth, bitmapHeight)
+        }
 
-        val bitmap = createBitmap(bitmapWidth, bitmapHeight, options.bitmapConfig.toSoftware())
-        val renderOptions = options.css?.let { RenderOptions().css(it) }
-        svg.renderToCanvas(Canvas(bitmap), renderOptions)
+        val svgRootWidth = svg.root?.width?.value ?: svgWidth
+        val svgRootHeight = svg.root?.height?.value ?: svgHeight
 
-        DecodeResult(
-            image = bitmap.toDrawable(options.context.resources).asCoilImage(),
+        val xScale: Float = bitmapWidth / svgRootWidth
+        val yScale: Float = bitmapHeight / svgRootHeight
+
+        val canvas = Canvas(bitmap)
+        canvas.scale(xScale, yScale)
+        svg.render(canvas)
+
+        return DecodeResult(
+            image = bitmap.asCoilImage(),
             isSampled = true, // SVGs can always be re-decoded at a higher resolution.
         )
     }
@@ -92,8 +96,7 @@ class SvgDecoder @JvmOverloads constructor(
             return dstWidth.toPx(scale) to dstHeight.toPx(scale)
         }
     }
-
-    class Factory @JvmOverloads constructor(
+    class Factory(
         val useViewBoundsAsIntrinsicSize: Boolean = true,
     ) : Decoder.Factory {
 
@@ -103,7 +106,7 @@ class SvgDecoder @JvmOverloads constructor(
             imageLoader: ImageLoader,
         ): Decoder? {
             if (!isApplicable(result)) return null
-            return SvgDecoder(result.source, options, useViewBoundsAsIntrinsicSize)
+            return SkiaSvgDecoder(result.source, options, useViewBoundsAsIntrinsicSize)
         }
 
         private fun isApplicable(result: SourceFetchResult): Boolean {
