+++ samples/shared/src/androidMain/kotlin/sample/common/imageLoaders.kt
@@ -4,7 +4,6 @@ import android.os.Build.VERSION.SDK_INT
 import coil3.ComponentRegistry
 import coil3.decode.GifDecoder
 import coil3.decode.ImageDecoderDecoder
-import coil3.decode.SvgDecoder
 import coil3.decode.VideoFrameDecoder
 
 internal actual fun ComponentRegistry.Builder.addPlatformComponents() {
@@ -14,8 +13,6 @@ internal actual fun ComponentRegistry.Builder.addPlatformComponents() {
     } else {
         add(GifDecoder.Factory())
     }
-    // SVGs
-    add(SvgDecoder.Factory())
     // Video frames
     add(VideoFrameDecoder.Factory())
 }
