+++ samples/shared/src/nonAndroidMain/kotlin/sample/common/imageLoaders.kt
@@ -1,8 +1,5 @@
 package sample.common
 
 import coil3.ComponentRegistry
-import coil3.decode.SvgDecoder
 
-internal actual fun ComponentRegistry.Builder.addPlatformComponents() {
-    add(SvgDecoder.Factory())
-}
+internal actual fun ComponentRegistry.Builder.addPlatformComponents() {}
